Initial Commit
Last version was loaded with bugs and given how quickly things were changing, it made no sense to keep pushing to SimpleUnityHelperScripts since breaking changes happened every other push.  This is a new repository to refresh the collection with all the previous changes that will absolutely explode any project that had the last version.  Many of the old structures will be removed and replaced.