Observable Extension Additions
Added -

MergeTriggersIntoIndex : Given a parameter array of Observables of the same type, returns a stream of indexes that signal which observable has emitted in time order.

TimerScan : Given an observable of type T, a "tick function" (triggers each timer increment), and a seed timer, returns a tuple stream of (value, timer) that completes after emitting the timer "on complete"

Added SwitchTo: simple extensions that selects a stream of streams and switches to the most recent one.

TakeWhile_IncludeLast: Given of stream of type T and a predicate of T -> bool, returns a stream that stops after the predicate returns false. (emits the value that fails the predicate then completes)